MemCardFile: Add some sanity checks for invalid file entries.
Clamp file sizes to the card's size, since this is not possible.
(GCN doesn't support sparse files.)

MemCardFile::size(): Return uint16_t, since files can be larger
than 255 blocks.

MemCardFilePrivate::size(): Get fatEntries.size() instead of
dirEntry->length in case the directory entry is invalid. Use
this function everywhere instead of accessing dirEntry->length.
that's larger than the memory card

+++ src/mcrecover/MemCardFile.hpp
@@ -126,10 +126,10 @@ class MemCardFile : public QObject
 		QString permissionAsString(void) const;
 
 		/**
-		 * Get the size, in blocks.
-		 * @return Size, in blocks.
+		 * Get the file size, in blocks.
+		 * @return File size, in blocks.
 		 */
-		uint8_t size(void) const;
+		uint16_t size(void) const;
 
 		/**
 		 * Get the banner image.
